requirements.txt
Flask==3.1.0
google_search_results==2.4.2
python-dotenv==1.0.1
Requests==2.32.3
serpapi==0.1.5
textblob==0.19.0
